# Async await analyze