… Sources/App/Commands/ReAnalyzeVersions.swift
… Sources/App/Controllers/API/API+PackageController.swift
… Sources/App/Core/AppMetrics.swift
… Tests/AppTests/AnalyzeErrorTests.swift
… Tests/AppTests/AnalyzerTests.swift
… Tests/AppTests/AnalyzerVersionThrottlingTests.swift
… Tests/AppTests/ErrorReportingTests.swift
… Tests/AppTests/MetricsTests.swift
… Tests/AppTests/PackageTests.swift
… Tests/AppTests/PipelineTests.swift
… Tests/AppTests/ReAnalyzeVersionsTests.swift
… Tests/AppTests/TwitterTests.swift