Integrating the S3 check into analysis finally pushed me over the edge making the a/a transition. The way we need to fetch versions and weave the S3 results in with the rest is just a nightmare within the ELF pipeline.

This is a big change, nonetheless. It went well and the tests highlighted a couple of areas that needed care (continuation on error).

One thing I still need to do is check timing behaviour. ELFs are eager and I need to make sure we don't run things serially now where they weren't before. I _think_ that's taken care of with the task group but I need to confirm timing.

- [x] confirm batch timing

We'll want to let this run on staging for a day or two before pushing this to prod.